backend/requirements.txt
fastapi
uvicorn
rawpy
Pillow
python-multipart
imageio
numpy
opencv-python
